Fix bug in MyGetBaseProjectDirectory
Neovim function system() captures both stdout and stderr while
we are interested only in stdout. This was the reason we failed
to detect when file is not in repository, as error message was
confused for the git root directory.

I took this opportunity to also fix the error reporting when
we ultimately fail to find the tag despite AutoTag's efforts.