Fix <CR> key behaviour in Orgmode
Note that the reason I am here using imap <expr> is to take advantage of
built-in autoindent mode. But, since I used normal command o to make a new line,
<CR> key lost the ability to split line at cursor position (when inside a list).
Thus I switch to inserting space at cursor position and then replacing it with
a new-line using r<CR> normal command. This worked fine to split a line, but it
again broke autoident when cursor was at the end of line (and there was no text
after it to carry it over to the new line). This required that I insert not just
space, but space + another character (i chose '_') that would be carried over to
the new line, triggering autoindent, and then deleting this superfluous character.

At the same time, we are also fixing another bug of not treating top level headline
(with a single star) as a headline!