Fix two bugs concerning Zetelkastten daily logs
One concerning navigation between Zetelkastten daily logs because
filename format is MM-DD-YYYY, so chronological and lexicographic
order are not the same when there are more then one year worth of
notes in directory; the remedy is to sort filenames chronologically
internally.

The other bug is failing to pad single digit months with zero when
forming Zetelkastten filename in ZetelkasttenCalendar_cb callback.