Orgmode: Work around the way nvim treats cursor position w.r.t conceal
While conceallevel>=2 hides concealed characters from being
displayed on the screen, everything else, including line wrapping
and horizontal scrolling of the window content remains unaffected,
acting as if all these characters are still there rather than
being governed according to what is visually displayed on the
screen and the visual cursor position as opposed to it's position
in the actual text contained in the buffer. This can lead to
inconsistencies like prematurely wrapping lines into unnecessary
many other lines, or, alternatively, scrolling window horizontally,
as if cursor left the screen, even though it is visually still
very far from the right edge, near the beggining of the line,
at the same time turning the screen into a sphagetti mess of
zigzag lines as each is shifted by a different amount depending
on how much concealed characters they contain, since rendering
code is unaware of concealment. In particular, this practicaly
renders useless using conceal to hide long URLs in Orgmode links!
See https://github.com/neovim/neovim/issues/14409.

In order to work around this, we introduce new preview mode
that continuously strips the URL parts from the links on the
current line (as they contribute the majority of concealed
characters) and then restores them as soon as you leave the
line; function FollowLinkUnderCursor is modified to look for
URLs in the stowed original version of the line.