Fix interaction between Preview mode and orgmode specific autoindent
Having 'o' and 'A' commands enter and leave Insert mode in
macros produced by MyOrgModeEnterKeyHandler() function triggers
the logic from Preview mode to respectively stow and restore
the text of the current line. This means that any changes to
this text we make in between leaving and then re-entering the
Insert mode would be undone, and this is exactly what happens
between the 'o' and 'A' commands in the aforementioned macros.

In this commit, the problem is solved by temporarily disabling
Preview mode for the duration of macro.