Preserve cursor position on Insert->Preview transition
We already dealt with keeping cursor in the same place when
stripping URLs from links in function PreviewModeUpdate(),
and this is the other side of the same coin. It is achieved
through InsertEnter autocmd by moving cursor in the approriate
position --- determined by an algorithm similar to that in
MyOrgModeGetLinkUnderCursor().

There is one more thing to take care of: it turns out us moving
cursor in InsertEnter interferes with operation of normal 'A'
command. The precise sequence of events is as follows:

   1. you initiate A command
   2. Vim moves cursor past the end of line
   3. this triggers CursorMoved event
   4. Vim prepares to enter Insert mode
   5. which triggers InsertEnter event
   6. finaly, user is in an Insert mode

Since URLs in the links are restored at 5, at 2. Vim still works
with smaller line length, and ultimately, cursor ends up in the
wrong place, somewhere in the middle of the line. To remedy this,
we remap 'A' to manually restore URLs, before calling builtin 'A'.
But beware, as these URLs might be stripped again at 3, or messed
up with in 5. afterwards, so we also take care of that too.
In particular, we keep b:PreviewModeStowedLine after intitally
restoring URLs to prevent them from being stripped again, but then
remember to still remove it before entering insert mode in order to
later pick up the changes --- that's what skip_next parameter is
all about.